Bugün Entity Framework kütüphanesini projeye dahil edip, bununla ilgii implementasyonaları yaptım. Northwind veritabanımıza Entity Framework aracılığıyla bağlandım ve burada CRUD operasyonlarımızı test edip eksiklerimi düzelttm, SOLID prensipleri ile kodlarımızı düzelte düzelte ilerliyoruz.

+++ Entities/Concrete/Product.cs
@@ -10,8 +10,8 @@ namespace Entities.Concrete
         public int ProductId { get; set; }
         public int CategoryId { get; set; }
         public string ProductName { get; set; }
-        public decimal UnitsInStock { get; set; }
-        public short UnitPrice { get; set; }
+        public short UnitsInStock { get; set; }
+        public decimal UnitPrice { get; set; }
 
     }
 }
